validation and currying
validate middleware to ensure it responds to .call

if an instance or proc are given as middleware and initialization args, then curry the args to use later for .call().  seems better than raising an error.  avoid using Method.currysince that will invoke the method if enough args are passed in